tests/wpt/web-platform-tests/workers/constructors/SharedWorker/Infinity-arguments.html

<!doctype html>
<title>Infinity as arguments</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  const worker = new SharedWorker(Infinity, Infinity);
  worker.port.onmessage = t.step_func_done(e => {
    assert_equals(e.data[0], 'Infinity', 'first arg (script name)');
    assert_equals(e.data[1], 'Infinity', 'second arg (worker name)');
  });
}, 'Test constructing a shared worker with Infinity');
</script>
